backend/membrane/pyproject.toml
[project]
name = "membrane"
version = "0.1.0"
description = "Membrane AgentCore"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-opentelemetry-distro>=0.14.2",
    "bedrock-agentcore>=1.2.0",
    "strands-agents>=1.23.0",   
]
